+++ bar-generif/bar-generif.properties
@@ -1,0 +1,8 @@
+compile.dependencies = intermine/objectstore/main, \
+           intermine/model/main, \
+           bio/core/main, \
+           intermine/integrate/main, \
+           bio/sources/generif/main
+
+have.file.custom.tgt = true
+converter.class = org.intermine.bio.dataconversion.BarGenerifConverter
