+++ bar-generif/src/main/resources/bar-generif_additions.xml
