+++ bar-generif/src/main/resources/bar-generif_keys.properties
