+++ bar-go-annotation/bar-go-annotation.properties
@@ -1,0 +1,7 @@
+compile.dependencies = intermine/objectstore/main,bio/core/main,\
+                       intermine/integrate/main, \
+                       bio/sources/go-annotation/main
+
+have.file.custom.tgt = true
+converter.class = org.intermine.bio.dataconversion.BarGoConverter
+postprocessor.class = org.intermine.bio.postprocess.BarGoPostprocess
