+++ bar-go-annotation/src/main/resources/bar-go-annotation_config.properties
