+++ bar-go-annotation/src/main/resources/bar-go-annotation_source.properties
