+++ bar-go-annotation/src/test/resources/GoConverterOboTest_src.txt
