+++ bar-go-annotation/src/test/resources/GoConverterOboTest_tgt.xml
