+++ bar-go-annotation/src/test/resources/go-tiny.obo
