+++ bar-go-annotation/src/test/resources/synbio_test
