--- go-annotation/go-annotation.properties
@@ -1,7 +1,0 @@
-compile.dependencies = intermine/objectstore/main,bio/core/main,\
-                       intermine/integrate/main, \
-                       bio/sources/go-annotation/main
-
-have.file.custom.tgt = true
-converter.class = org.intermine.bio.dataconversion.GoConverter
-postprocessor.class = org.intermine.bio.postprocess.GoPostprocess
