--- po-annotation/.checkstyle
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<fileset-config file-format-version="1.2.0" simple-config="false">
-    <fileset name="main-java-only" enabled="true" check-config-name="FlyMine" local="false">
-        <file-match-pattern match-pattern=".*java$" include-pattern="true"/>
-        <file-match-pattern match-pattern=".*build/" include-pattern="false"/>
-        <file-match-pattern match-pattern=".*Test\.java$" include-pattern="false"/>
-    </fileset>
-</fileset-config>
