--- po-annotation/.classpath
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="main/src"/>
-	<classpathentry kind="src" path="test/src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/bio-testall-dbmodel"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/intermine-integrate-main"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/intermine-objectstore-main"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/bio-core-main"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/intermine-objectstore-test"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/intermine-integrate-test"/>
-	<classpathentry kind="lib" path="test/resources"/>
-	<classpathentry kind="lib" path="main/resources"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/bio-postprocess-main"/>
-	<classpathentry kind="output" path="bin"/><classpathentry combineaccessrules="false" kind="src" path="/intermine-model-test"/><classpathentry combineaccessrules="false" kind="src" path="/intermine-model-main"/>
-</classpath>
