--- po-annotation/build.gradle
@@ -1,29 +1,0 @@
-sourceSets {
-    main {
-        java {
-            srcDirs = ['src/main/java', 'build/gen']
-        }
-        resources {
-            srcDirs = ['src/main/resources']
-        }
-    }
-    test {
-        java {
-            srcDirs = ['src/test/java']
-        }
-        resources {
-            srcDirs = ['src/test/resources']
-        }
-    }
-}
-
-dependencies {
-    compile group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false //to read genomic_keyDefs.properties
-    compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
-    runtime fileTree(dir: 'libs', include: '*.jar') //antlr-2.7.6-caching.jar
-    bioModel group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false
-}
-
-processResources {
-    from('.') { include ("*.properties")}
-}
