--- po-annotation/po-annotation.properties
@@ -1,6 +1,0 @@
-compile.dependencies = intermine/objectstore/main,bio/core/main,\
-                       intermine/integrate/main, \
-                       bio/sources/po-annotation/main
-have.file.custom.tgt = true
-converter.class = org.intermine.bio.dataconversion.PoConverter
-postprocessor.class = org.intermine.bio.postprocess.PoPostprocess
