--- po-annotation/src/main/resources/po-annotation_additions.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<classes>
-<class name="POAnnotation" extends="OntologyAnnotation" is-interface="true">
-  <attribute name="annotationExtension" type="java.lang.String"/>
-</class>
-<class name="Gene" is-interface="true">
-  <collection name="poAnnotation" referenced-type="POAnnotation"/>
-</class>
-<class name="POEvidenceCode" extends="OntologyAnnotationEvidenceCode" is-interface="true"/>
-<class name="POTerm" extends="OntologyTerm" is-interface="true"/>
-<class name="POEvidence" extends="OntologyEvidence" is-interface="true">
-  <attribute name="withText" type="java.lang.String"/>
-  <collection name="with" referenced-type="BioEntity"/>
-</class>
-<class name="DatabaseReference" is-interface="true">
-  <attribute name="identifier" type="java.lang.String"/>
-  <reference name="source" referenced-type="DataSource"/>
-  <reference name="subject" referenced-type="Publication" reverse-reference="crossReferences" />
-</class>
-<class name="Publication" is-interface="true">
-  <collection name="crossReferences" referenced-type="DatabaseReference" reverse-reference="subject"/>
-</class>
-</classes>
