--- po-annotation/src/main/resources/po-annotation_config.properties
@@ -1,46 +1,0 @@
-# For each taxon id specify the field name of Gene that should be filled
-# in.  Should be either 'symbol' or 'primaryIdentifier' (default)
-# `typeAnnotated` - the type of object to be annotated, either protein or gene.  gene is the default.
-# http://www.geneontology.org/GO.format.gaf-2_0.shtml
-# readcolumn - which column to use for identifier - identifier or symbol
-
-# Uniprot annotations
-7165.typeAnnotated=protein
-7165.identifier=primaryAccession
-7460.typeAnnotated=protein
-7460.identifier=primaryAccession
-9606.typeAnnotated=protein
-9606.identifier=primaryAccession
-9615.typeAnnotated=protein
-9615.identifier=primaryAccession
-
-# Mouse
-#10090.identifier=secondaryIdentifier
-#10090.typeAnnotated=gene
-
-# Arabidopsis
-3702.identifier=secondaryIdentifier
-3702.readColumn=identifier
-3702.typeAnnotated=gene
-
-226900.typeAnnotated=protein
-226900.identifier=primaryAccession
-
-1392.typeAnnotated=protein
-1392.identifier=primaryAccession
-
-83333.typeAnnotated=protein
-83333.identifier=primaryAccession
-
-634956.typeAnnotated=protein
-634956.identifier=primaryAccession
-
-568206.typeAnnotated=protein
-568206.identifier=primaryAccession
-
-592021.typeAnnotated=protein
-592021.identifier=primaryAccession
-
-224308.typeAnnotated=protein
-224308.identifier=primaryAccession
-
