--- po-annotation/src/main/resources/po-annotation_keys.properties
@@ -1,12 +1,0 @@
-Gene.key_primaryidentifier=primaryIdentifier
-Gene.key_secondaryidentifier_org=secondaryIdentifier, organism
-#Gene.key_symbol_org=symbol, organism
-Protein.key_primaryacc=primaryAccession
-Organism.key_taxonid=taxonId
-OntologyTerm.key_identifier=identifier
-OntologyTerm.key_name_ontology=name, ontology
-Ontology.key_name=name
-POEvidenceCode.key_code=code
-Publication.key_pubmed=pubMedId
-DataSet.key_name=name
-DataSource.key_name=name
