--- po-annotation/src/main/resources/po-annotation_source.properties
@@ -1,3 +1,0 @@
-have.file.custom.tgt = true
-converter.class = org.intermine.bio.dataconversion.PoConverter
-postprocessor.class = org.intermine.bio.postprocess.PoPostprocess
