--- po-annotation/src/main/resources/po-evidence-codes
@@ -1,5 +1,0 @@
-IDA	Inferred from Direct Assay	http://geneontology.org/page/ida-inferred-direct-assay
-IEP	Inferred from Expression Pattern 	http://geneontology.org/page/iep-inferred-expression-pattern
-IMP	Inferred from Mutant Phenotype	http://geneontology.org/page/imp-inferred-mutant-phenotype
-NAS	Non-traceable Author Statement	http://geneontology.org/page/nas-non-traceable-author-statement
-TAS	Traceable Author Statement	http://geneontology.org/page/tas-traceable-author-statement
