--- po-annotation/src/test/build.xml
@@ -1,38 +1,0 @@
-<project name="sources-go-annotation-test" default="default" basedir="."
-         xmlns:imtestsproj="http://www.intermine.org/ns/im-tests-proj/1"
-         xmlns:imdbproj="http://www.intermine.org/ns/im-task-proj/1">
-
-  <description>build and test sources-po-annotation-test</description>
-
-  <import file="../../../../imbuild/tests.xml"/>
-  <import file="../../../../imbuild/task.xml"/>
-
-  <target name="-pre-test" depends="build-db.bio-fulldata-test"/>
-
-  <target name="do-test" depends="-init-macrodef-junit">
-    <imtestsproj:junit/>
-  </target>
-
-  <target name="-init-build-db" depends="init, -init-deps, -init-task-xml"/>
-
-  <target name="-define-task-classpath">
-    <!-- Override to alter the classpath created -->
-    <path id="task.class.path">
-      <pathelement location="${build.task.dir}"/>
-      <pathelement location="${build.test.resources.dir}"/>
-      <pathelement location="${resources.dir}"/>
-      <path refid="project.execute.path"/>
-    </path>
-  </target>
-
-  <target name="build-db.bio-fulldata-test" depends="-init-build-db">
-    <imdbproj:build-db osname="os.bio-fulldata-test"
-                       model="fulldata"/>
-  </target>
-
-  <!-- index the fulldata test database -->
-  <target name="create-indexes-genomic-fulldata-test" depends="-init-build-db">
-    <create-indexes alias="os.bio-fulldata-test" attributeIndexes="false"/>
-  </target>
-
-</project>
