--- po-annotation/src/test/project.properties
@@ -1,15 +1,0 @@
-compile.dependencies = bio/sources/po-annotation/main,\
-                       intermine/integrate/test,\
-                       bio/test-all/dbmodel,\
-                       bio/core/main
-
-# choose the intermine.properties file from $HOME:
-intermine.properties.file = intermine-bio-test.properties
-
-default.intermine.properties.file = ../../../default.intermine.test.properties
-
-# classpath to instrument with emma
-instrument.path=../main/dist/intermine-objectstore.jar
-instrument.src.path=../main/src/
-
-extra.project.dependencies = bio/test-all/dbmodel
