--- po/.checkstyle
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<fileset-config file-format-version="1.2.0" simple-config="false">
-  <fileset name="all" enabled="true" check-config-name="FlyMine" local="false">
-    <file-match-pattern match-pattern="." include-pattern="true"/>
-  </fileset>
-</fileset-config>
