--- po/build.gradle
@@ -1,20 +1,0 @@
-sourceSets {
-    main {
-        java {
-            srcDirs = ['build/gen']
-        }
-        resources {
-            srcDirs = ['resources']
-        }
-    }
-}
-
-dependencies {
-    compile group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false  //to read genomic_keyDefs.properties
-    compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
-    bioModel group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false
-}
-
-processResources {
-    from('.') { include ("*.properties")}
-}
