--- po/po.properties
@@ -1,6 +1,0 @@
-have.file.obo = true
-obo.ontology.name=PO
-obo.ontology.url=http://www.plantontology.org
-obo.term.class=POTerm
-
-
