--- po/resources/po_additions.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0"?>
-<classes>
-<class name="POTerm" extends="OntologyTerm" is-interface="true"/>
-</classes>
