--- po/resources/po_keys.properties
@@ -1,5 +1,0 @@
-OntologyTerm.key_identifier=identifier
-OntologyTerm.key_name_ontology=name, ontology
-Ontology.key_title=name
-DataSet.key_title=name
-DataSource.key_name=name
