+++ settings.gradle
@@ -8,12 +8,11 @@ include ':bio-source-bar-tair-gff'
 project(':bio-source-bar-tair-gff').projectDir = new File(settingsDir, './bar-tair-gff')
 
 // This is mostly from upstream Intermine
-// Only ath changes added
-include ':bio-source-go-annotation'
-project(':bio-source-go-annotation').projectDir = new File(settingsDir, './go-annotation')
+// This might not work upstream
+include ':bio-source-bar-go-annotation'
+project(':bio-source-bar-go-annotation').projectDir = new File(settingsDir, './bar-go-annotation')
 
 // This is from Araport fork of Intermine.
-// This needs update to Intermine v4.
 include ':bio-source-uniprot'
 project(':bio-source-uniprot').projectDir = new File(settingsDir, './uniprot')
 
@@ -29,7 +28,7 @@ project(':bio-source-psi').projectDir = new File(settingsDir, './psi')
 
 // ID resolver cant be skipped 
 // Overwrite resolved IDs
-include ':bio-source-generif'
-project(':bio-source-generif').projectDir = new File(settingsDir, './generif')
+include ':bio-source-bar-generif'
+project(':bio-source-bar-generif').projectDir = new File(settingsDir, './bar-generif')
